restaurant_ordering_backend/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
psycopg2-binary==2.9.1
django-cors-headers==3.7.0
djangorestframework-simplejwt==5.1.0
